+++ test/physics/test_scattering.jl
@@ -1,14 +1,3 @@
-@testitem "ScatteringData construction" begin
-    grid = PitchAngleGrid(180:-30:0)
-    sd = AURORA.ScatteringData(grid)
-
-    @test sd isa AURORA.ScatteringData
-    @test size(sd.Ω_subbeam_relative, 1) == grid.n_beams
-    @test size(sd.P_scatter, 1) == size(sd.P_scatter, 2)
-    @test length(sd.Ω_beam) == grid.n_beams
-    @test length(sd.θ_scatter) > 0
-end
-
 @testitem "Rotation matrices" begin
     # Define some parameters for the tests
     θ_lims = 180:-10:0